Optimizations and fixes
Using float versions of math functions, inlined normal axis function.  Changed tonemapping from modified Reinhard to ACES and added gamma correction.  Fixed hdri horizontal sampling.  Included 1080p 1 sample render with refractive cubes.

+++ vector3.c
@@ -20,7 +20,7 @@ float vec3_mag2(Vector3 *a) {
 }
 
 float vec3_mag(Vector3 *a) {
-    return sqrt(vec3_mag2(a));
+    return sqrtf(vec3_mag2(a));
 }
 
 float vec3_dot(Vector3 *a, Vector3 *b) {
